change logout button
