+++ src/components/Menu/MenuBar.module.scss
@@ -81,6 +81,14 @@
                 font-size: 1rem;
                 color: inherit;
             }
+            .logout {
+                margin-right: 10px;
+                font-size: 1rem;
+                color: red;
+            }
+        }
+        .menuItemLogout {
+            color: red;
         }
 
     }
